requirements.txt
pysc2==3.0.0
numpy
pandas
pygame
absl-py
protobuf<=3.20.3
